backend/ai/gemini_client.py
"""
Gemini API helper — one simple function to generate text.
"""

import os
from google import genai
from dotenv import load_dotenv

load_dotenv()

client = genai.Client(api_key=os.getenv("GEMINI_API_KEY"))

MODEL = "gemini-2.5-flash"


def generate_text(prompt: str) -> str:
    """Send a prompt to Gemini and return the text response."""
    response = client.models.generate_content(
        model=MODEL,
        contents=prompt,
    )
    return response.text
