backend/requirements.txt
fastapi
uvicorn
asyncpg
python-dotenv
requests
google-genai
